+++ src/lantern/stores/gitlab_cache.py
@@ -137,15 +137,25 @@ class GitLabLocalCache:
         self._token = gitlab_token
         self._source_ = gitlab_source
         self._frozen = frozen
-
         self._cache_path = path / "cache.db"
-        self._flash: dict[str, RecordRevision] = {}
 
-    @cached_property
+        self._flash: dict[str, RecordRevision] = {}
+        self._conn: Engine | None = None
+
+    def __getstate__(self):  # noqa: ANN204
+        """Close and unset engine to allow pickling."""
+        if self._conn is not None:
+            self._engine.disconnect_all()
+            self._conn = None
+        return self.__dict__.copy()
+
+    @property
     def _engine(self) -> Engine:
         """Engine for backing database."""
-        self._logger.info(f"Connecting to SQLite database at: '{self._cache_path.resolve()}'")
-        return Engine.from_uri(f"sqlite://{self._cache_path.resolve()}")
+        if self._conn is None:
+            self._logger.info(f"Connecting to SQLite database at: '{self._cache_path.resolve()}'")
+            self._conn = Engine.from_uri(f"sqlite://{self._cache_path.resolve()}")
+        return self._conn
 
     @cached_property
     def _online(self) -> bool:
@@ -555,7 +565,7 @@ class GitLabLocalCache:
             and len(file_identifiers) > 0
             and all(fid in self._flash for fid in file_identifiers)
         ):
-            self._logger.info(f"Loading {len(file_identifiers)} records from flash")
+            self._logger.debug(f"Loading {len(file_identifiers)} records from flash")
             return [self._flash[fid] for fid in file_identifiers]
 
         query = SQL("SELECT record_pickled FROM record")