+++ tasks/css.py
@@ -7,7 +7,6 @@ from tempfile import TemporaryDirectory
 from boto3 import client as S3Client  # noqa: N812
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 from moto import mock_aws
-from tests.conftest import _init_fake_store
 from tests.resources.records.admin_keys.testing_keys import load_keys
 from tests.resources.stores.fake_records_store import FakeRecordsStore
 
@@ -75,7 +74,7 @@ def export_test_site(export_path: Path) -> None:
         meta=meta,
         s3=s3_client,
         logger=logger,
-        init_store=_init_fake_store,
+        store=store,
         selected_identifiers={record.file_identifier for record in store.select()},
     )
     exporter.export()